requirements.txt
requests==2.31.0
folium==0.14.0
python-dotenv==1.0.0
pandas==2.1.4
geopy==2.4.1
